client.http.sync: Drop use of requests `Session`
Ordinarily, this session object is responsible for doing things like
managing and expiring cookies, which normally is a desirable thing for
HTTP clients.

SkySpark gets horribly confused when you do this.  It will demand an
attestation key if it sees a cookie round-tripped by the HTTP client.

We actually round-trip things like cookies at a higher level, so it's
redundant for `requests` to do it for us, and plus it's harmful in this
case, we actually don't want to store the cookie in this situation.